Resolve editor state name conflicts during serialization (#344)
Fixes an issue where controls with the same case-insensitive names can conflict on the file system when generating editorstate.json files for the control. This uses the existing name collision handling logic to generate a new filename when needed. Changes the way the TopParentName is loaded into controls on deserialization to improve accuracy.