Directly pass through custom page schema files, and handle any validation in canvas (#439)
## Problem

Related to https://github.com/microsoft/Power-Fx/pull/663, we're consolidating the type definition logic for this, and using yaml as the native format for it in the .msapp. As such, the .PALT tool does not need to do anything special to handle this file. 

## Solution
Removes special handling for schema.yaml

## Changes
Removes special handling for schema.yaml, associated POCOs, and passes the value through as text. 

## Validation
Testing integration with related changes is ongoing and this will not be checked in until all are ready.